test: move mockSocialServer to separate file
