+++ test/unit/friendships.spec.ts
@@ -1,6 +1,6 @@
 import { expect } from 'chai'
-import jsonServer from 'json-server'
 import { getFriendsFromSocialService } from '../../src/FriendsManagementClient'
+import { mockSocialServer } from './socialServerMock'
 
 describe('friendships from social server', () => {
     const PORT = 3131
@@ -42,42 +42,3 @@ describe('friendships from social server', () => {
         })
     })
 })
-
-function mockSocialServer(userId: string, validToken: string, anotherUserId: string, PORT: number) {
-    const server = jsonServer.create()
-
-    // No friends userId
-    server.get(`/v1/friendships/${userId}`, (req, res) => {
-        if (isValidToken(req)) {
-            res.jsonp({
-                friendships: []
-            })
-        } else {
-            res.sendStatus(401)
-        }
-    })
-
-    // Two friends anotherUserId
-    server.get(`/v1/friendships/${anotherUserId}`, (req, res) => {
-        if (isValidToken(req)) {
-            res.jsonp({
-                friendships: [
-                    {
-                        address: '0xc0ffee254729296a45a3885639AC7E10F9d54979'
-                    },
-                    {
-                        address: '0x86F842D7Ea37EbEC6248eF1652E7DB971C631CCC'
-                    }
-                ]
-            })
-        } else {
-            res.sendStatus(401)
-        }
-    })
-
-    server.listen(PORT, () => console.log(`JSON Server is running on port ${PORT}`))
-
-    function isValidToken(req) {
-        return req.headers && req.headers['authorization'] === `Bearer ${validToken}`
-    }
-}
